requirements.txt
streamlit
chromadb
faiss-cpu
numpy
pandas
requests
python-dotenv
openai
groq